기존 회원용 화면 + 타입 수정

+++ src/main/java/com/example/keyosk_backend/model/Item.java
@@ -12,7 +12,7 @@ public class Item {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ITEM_ID")
-    private int itemId;
+    private Long itemId;
 
     private String itemType;
